Cleaned up the code and added some annotations.
Moved alpha level for shaded statistically significant regions to be defined in the plotting setting defaults and manual plotting scripts, rather than it be hard-coded in the plotting functions. This allow users to adjust the alpha level of the shaded regions.

Added housekeeping clear/close all commands to start of manual plotting scripts, and added some variables that were not defined in these scripts.